refactor: decouple reference harness from main cli
